tests/LoadBalancer.Tests/LoadBalancerServiceTests.cs
using Moq;
using System;
using System.Collections.Generic;
using System.Linq;
using System.Net.Http;
using System.Text;
using System.Threading.Tasks;

namespace LoadBalancer.Tests;

public class LoadBalancerServiceTests
{
    [Fact]
    public async Task ProcessRequestAsync_SelectsHealthyServer()
    {
        // Arrange
        var backendServers = new List<BackendServer>
        {
            new BackendServer("Server1", "http://localhost:5001", "/health"),
            new BackendServer("Server2", "http://localhost:5002", "/health"),
        };

        // Mock ILoadBalancingAlgorithm and IHealthCheck
        var loadBalancingAlgorithmMock = new Mock<ILoadBalancingAlgorithm>();
        var healthCheckMock = new Mock<IHealthCheck>();

        // Set up mock to always return the first server as healthy
        healthCheckMock.Setup(h => h.GetHealthyServersAsync(It.IsAny<List<BackendServer>>()))
            .ReturnsAsync(new List<BackendServer> { backendServers[0] });

        var loadBalancer = new LoadBalancerService(backendServers, loadBalancingAlgorithmMock.Object, healthCheckMock.Object);

        // Act
        var response = await loadBalancer.ProcessRequestAsync();

        // Assert
        // Ensure that the response is not empty (actual assertions would depend on your application logic)
        Assert.NotNull(response);

        // Verify that GetHealthyServersAsync was called
        healthCheckMock.Verify(h => h.GetHealthyServersAsync(It.IsAny<List<BackendServer>>()), Times.Once);
    }

    [Fact]
    public async Task ProcessRequestAsync_NoHealthyServerAvailable()
    {
        // Arrange
        var backendServers = new List<BackendServer>
        {
            new BackendServer("Server1", "http://192.168.1.153:80", "/health"),
            new BackendServer("Server2", "http://192.168.1.153:3000", "/health"),
            new BackendServer("Server3", "http://192.168.1.153:3001", "/health"),
        };

        // Mock ILoadBalancingAlgorithm and IHealthCheck
        var loadBalancingAlgorithmMock = new Mock<ILoadBalancingAlgorithm>();
        var healthCheckMock = new Mock<IHealthCheck>();

        // Set up mock to always return an empty list (no healthy servers)
        healthCheckMock.Setup(h => h.GetHealthyServersAsync(It.IsAny<List<BackendServer>>()))
            .ReturnsAsync(new List<BackendServer>());

        var loadBalancer = new LoadBalancerService(backendServers, loadBalancingAlgorithmMock.Object, healthCheckMock.Object);

        // Act
        var response = await loadBalancer.ProcessRequestAsync();

        // Assert
        // Ensure that the response indicates no healthy server available (actual assertions would depend on your application logic)
        Assert.Equal("No healthy backend server available.", response);

        // Verify that GetHealthyServersAsync was called
        healthCheckMock.Verify(h => h.GetHealthyServersAsync(It.IsAny<List<BackendServer>>()), Times.Once);
    }

    // Add more test methods to cover other scenarios, load balancing algorithms, etc.
}
